Project_2/fitness-flask-app/requirements.txt
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
mysqlclient==2.0.3